feat: add import command and download file to import products

+++ database/migrations/2024_11_02_003801_create_products_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->timestamp('imported_t')->nullable();
             $table->string('url');
             $table->string('creator');
-            $table->bigInteger('created_t');
-            $table->bigInteger('last_modified_t');
+            $table->unsignedInteger('created_t');
+            $table->unsignedInteger('last_modified_t');
             $table->string('product_name');
             $table->string('quantity');
             $table->string('brands');
